zan-samaki-backend/requirements.txt
Django==5.1
djangorestframework==3.15.2
djoser==2.3.0
djangorestframework-simplejwt==5.3.1
Pillow==10.4.0
django-cors-headers==4.4.0
channels==4.1.0
channels-redis==4.2.0
python-decouple==3.8
psycopg2-binary==2.9.9
black==24.4.2

